// src/app/model/ministry-customerget.model.ts
export interface MinistryCustomerGetModel {
    customerNumber: string| null;
    ministryName: string | null;
    customerAddress: string | null;
    customerName: string | null;
    zoneName: string| null;
    circleName: string| null;
    centerName: string| null;
    locationCode: string| null;
    locationDesc: string| null;
    religiousCode: string | null;
    
}